update longe to integer

+++ src/main/java/com/trego/dao/entity/Medicine.java
@@ -12,7 +12,7 @@ public class Medicine {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "vendor_medicine_id")
-    private Long vendorMedicineId;
+    private Integer vendorMedicineId;
 
     @Column(columnDefinition = "TEXT")
     private String name;
@@ -73,11 +73,11 @@ public class Medicine {
 
     // Delegate getId()/setId() to vendorMedicineId (the actual PK)
     // because the legacy "id" column is NULL in DB and has been removed
-    public Long getId() {
+    public Integer getId() {
         return this.vendorMedicineId;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.vendorMedicineId = id;
     }
 }
